% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSensitiveGene.R
\name{GetSensitivegene}
\alias{GetSensitivegene}
\title{Screening for sensitive genes}
\usage{
GetSensitivegene(object, min_nClusters = "Default", min_nCell = 0, HVG_Anno)
}
\arguments{
\item{object:}{The seurat object, with HVG_Statistic finished.}

\item{min_nClusters:}{nClusters of sensitive genes should be detect as HVGs at least,default:"Default"(Half of nCluster)}

\item{min_nCell:}{nCells of sensitive genes should be detect as HVGs as least,default:0}

\item{HVG_Anno:}{Statistics result on the high variable genes in each subclusterd}
}
\description{
Screening for sensitive genes based on the CV Ranking and Shannon Index
}
