% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HVG_Statistic.R
\name{HVG_Statistic}
\alias{HVG_Statistic}
\title{Statistics on high variable genes of each subclusters}
\usage{
HVG_Statistic(
  object,
  First_time_unsupervised_clustering_label = "First_time_unsupervised_clustering"
)
}
\arguments{
\item{object:}{The seurat object, produced by Seurat_Pipline Function or provided by User}

\item{First_time_unsupervised_clustering_label:}{The labels of Unsupervised clustering results}
}
\description{
Statistics on the high variable genes in each subclusterd, divided by the First-time unsupervised clustering

Only need to provide the label of First-time unsupervised clustering.
}
