% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReSelectVariableFeatures.R
\name{ReSelectVariableFeatures}
\alias{ReSelectVariableFeatures}
\title{Remove sensitive genes from variable Features}
\usage{
ReSelectVariableFeatures(object, SensitiveGene)
}
\arguments{
\item{object:}{Seurat object, with HVG_Statistic and GetSensitiveGene finished}

\item{SensitiveGene:}{SensitiveGene List of the scRNA-seq data. The sensitive gene can be get by running "GetSensitiveGene"}
}
\description{
Only removing the sensitive genes from the variable features, not transcriptome. In this way, the unsupervised clustering result will be improved with transcript no change.
}
