% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_Pipline.R
\name{Seurat_Pipline}
\alias{Seurat_Pipline}
\title{The default pipline for processing scRNA-seq by Seurat}
\usage{
Seurat_Pipline(
  data.dir,
  PC = 40,
  resolution = 0.6,
  mt.cut_off = 20,
  min_nFeature.cut_off = 200,
  sample_name = "scRNA-seq",
  data_type = "Expression_matrix",
  filter_doublet = TRUE
)
}
\arguments{
\item{data.dir:}{Path to the matrix from CellRanger reults, such as "~/sample_CellRanger_result/outs/filtered_feature_bc_matrix"}

\item{PC:}{Number of PC used for demension reduction and clustering. default:40}

\item{resolution:}{Choose appropriate resolution value to adjust the number of clusters. default:0.6}

\item{mt.cut_off:}{Filter out the cells with a high percentage of MT-genes. default:20(\%) (Only apply to samples for human)}

\item{min_nFeature.cut_off:}{Filter low quality cell by number of Features in each cells. default:200}

\item{sample_name:}{Samples name of scRNA-seq data. default:"scRNA-seq"}

\item{data_type:}{Type of scRNA-seq data. Temporary, only "Expression_matrix" and "TotalSeq" data are supported}

\item{filter_doublet:}{Whether use DoubletFinder to remove Doublet in scRNA-seq; Attention, DoubletFinder performed not pretty well in detect the doublet of same type of cells.}
}
\description{
Analysis the single cell data according to the default parameters of Seurat. Users can adjust the parameters according to their needs.

This step can be skip, if user can provided the processed data.

The processed data should finish several steps by using Seurat: Normalized data, FindVariableFeatures, ScaleData, RunPCA, FindNeighbors, FindClusters and RunUMAP.
}
