% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReClustering.R
\name{ReClustering}
\alias{ReClustering}
\title{Reclustering after removing sensitive genes}
\usage{
ReClustering(object, PC = 40, resolution = 0.6)
}
\arguments{
\item{object:}{Seurat object, with HVG_Statistic, GetSensitiveGene and ReSelectVariableFeatures finished}

\item{PC:}{Number of PC used for demension reduction and clustering. default:40}

\item{resolution:}{Choose appropriate resolution value to adjust the number of clusters. default:0.6}
}
\description{
After removing the sensitive gene, reclusering the single-cell data.
}
